+++ service/monitor/v1/events/api_op_page.go
@@ -20,7 +20,7 @@ type EventsPageOptions struct {
 	ActorSid        *string
 	EventType       *string
 	ResourceSid     *string
-	SourceIPAddress *string
+	SourceIpAddress *string
 	StartDate       *time.Time
 	EndDate         *time.Time
 }
@@ -37,7 +37,7 @@ type PageEventResponse struct {
 	ResourceType    string                 `json:"resource_type"`
 	Sid             string                 `json:"sid"`
 	Source          string                 `json:"source"`
-	SourceIPAddress string                 `json:"source_ip_address"`
+	SourceIpAddress string                 `json:"source_ip_address"`
 	URL             string                 `json:"url"`
 }